+++ htdocs/js/features.js
@@ -1,6 +1,6 @@
 "use strict";
 // SPDX-License-Identifier: GPL-3.0-or-later
-// myMPD (c) 2018-2024 Juergen Mang <mail@jcgames.de>
+// myMPD (c) 2018-2025 Juergen Mang <mail@jcgames.de>
 // https://github.com/jcorporation/mympd
 
 /** @module features_js */
@@ -33,6 +33,10 @@ function setFeatures() {
     features.viewTitle = settings.webuiSettings.viewTitles;
     //stickers config value
     features.featStickersEnabled = settings.features.featStickersEnabled;
+
+    features.featPagination = settings.webuiSettings.endlessScroll === false
+        ? true
+        : false;
 
     //mpd features
     if (settings.partition.mpdConnected === true) {