+++ src/web_server/utility.h
@@ -1,6 +1,6 @@
 /*
  SPDX-License-Identifier: GPL-3.0-or-later
- myMPD (c) 2018-2024 Juergen Mang <mail@jcgames.de>
+ myMPD (c) 2018-2025 Juergen Mang <mail@jcgames.de>
  https://github.com/jcorporation/mympd
 */
 
@@ -73,8 +73,9 @@ bool check_imagescache(struct mg_connection *nc, struct mg_http_message *hm,
 sds webserver_find_image_file(sds basefilename);
 bool find_image_in_folder(sds *coverfile, sds music_directory, sds path, sds *names, int names_len);
 void webserver_send_error(struct mg_connection *nc, int code, const char *msg);
-void webserver_serve_file(struct mg_connection *nc, struct mg_http_message *hm, const char *path, const char *file);
-void webserver_send_header_ok(struct mg_connection *nc, size_t len, const char *extra_headers);
+void webserver_serve_file(struct mg_connection *nc, struct mg_http_message *hm,
+        const char *headers, const char *file);
+void webserver_send_header_ok(struct mg_connection *nc, size_t len, const char *headers);
 void webserver_send_header_redirect(struct mg_connection *nc, const char *location, const char *headers);
 void webserver_send_header_found(struct mg_connection *nc, const char *location, const char *headers);
 void webserver_send_cors_reply(struct mg_connection *nc);